// src/canister/icpswap/icpswap.did.d.ts
export interface PublicTokenOverview {
    id: bigint;
    volumeUSD1d: number;
    volumeUSD7d: number;
    totalVolumeUSD: number;
    name: string;
    volumeUSD: number;
    feesUSD: number;
    priceUSDChange: number;
    address: string;
    txCount: bigint;
    priceUSD: number;
    standard: string;
    symbol: string;
}